OSEmu
Here is a sample for such a simple emu. It includes the basic needed
Camd35 UDP functionality and is really very simple and compact. It
should need almost no ressources.
The real emu part is not included. This needs to be called in OSEmu.c
at "HANDLE ECM HERE!" but that should be rather easy to achieve.
Probably the code that was floating around here should work with only
some small adaptations...

Build: Just enter "make", there are no options necessary as long as you
don't crosscompile (but as it's standard c code with no dependencies it
should work quite easy on all plattforms)
Start: Start the built OSEmu binary. The necessary options will get
displayed as usage info (a single user account and the port). It
automatically listens on all IP addresses on the machine, including
localhost.
Use in OSCam: Create a camd35 reader with caid/ident defined
accordingly to what you emulate and maybe also services if you want.
Use the user/password/port that you specified on start of OSEmu and it
should start to get decoded.